python_code/api/requirements.txt
pandas>=2.1.0
python-dotenv==1.0.1
openai==1.50.2
runpod==1.7.1
pinecone==5.3.1
httpx==0.27.0
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers>=3.0.0
